requirements.txt
# Core
opencv-python>=4.8
mediapipe>=0.10
pyautogui>=0.9.54
face_recognition>=1.3

# System tray (optional but recommended)
pystray>=0.19
Pillow>=10.0

# Input monitoring — for bypass key detection (alternative to tkinter bind)
keyboard>=0.13

# Startup helper (Windows only — optional)
# pywin32>=306
